cc: refactor DeepPot to support multiple backends
