+++ source/api_cc/include/common.h
@@ -19,6 +19,8 @@
 namespace deepmd {
 
 typedef double ENERGYTYPE;
+// TODO: currently we only implement TF; reserve for future use
+enum Backend { TENSORFLOW, PYTORCH, PADDLE, UNKNOWN };
 
 struct NeighborListData {
   /// Array stores the core region atom's index
